corrections after review

+++ afs-core/src/main/java/com/powsybl/afs/AbstractNodeBase.java
@@ -1,4 +1,4 @@
-/*
+/**
   Copyright (c) 2017, RTE (http://www.rte-france.com)
   This Source Code Form is subject to the terms of the Mozilla Public
   License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -15,7 +15,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.io.UncheckedIOException;
+import java.nio.file.FileAlreadyExistsException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.io.File;
@@ -33,8 +34,6 @@ import java.util.Optional;
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
 public abstract class AbstractNodeBase<F> {
-
-    private static final long MIN_DISK_SPACE_THRESHOLD = 10;
 
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractNodeBase.class);
 
@@ -185,47 +184,44 @@ public abstract class AbstractNodeBase<F> {
         return childNodes.stream().filter(nodeInfo -> !nodeInfo.getId().equals(getId())).anyMatch(nodeInfo -> nodeInfo.getName().equals(name));
     }
 
-    public void archive(Path dir, boolean useZip){
-
-        //if it's a copy paste, this test is already done
-        if(useZip) {
-            File archiveFile = dir.toFile();
-            long freeSpacePercent = 100 * archiveFile.getFreeSpace() / archiveFile.getTotalSpace();
-            LOGGER.info("Copying into drive with {}% free space", freeSpacePercent);
-            if (freeSpacePercent < MIN_DISK_SPACE_THRESHOLD) {
-                throw new AfsException("Not enough space");
-            }
-        }
+    public void archive(Path dir, boolean useZip) throws IOException {
 
         Objects.requireNonNull(dir);
-        try {
-            new AppStorageArchive(storage).archive(info, dir);
-            if(useZip){
-                Path zipPath = dir.resolve(info.getId() + ".zip");
-                Utils.zip(dir.resolve(info.getId()), zipPath);
-            }
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
+
+        if (Files.exists(dir.resolve(info.getId() + ".zip"))) {
+            throw new FileAlreadyExistsException("Archive already exist");
+        }
+
+        new AppStorageArchive(storage).archive(info, dir);
+        if (useZip) {
+            Path zipPath = dir.resolve(info.getId() + ".zip");
+            Utils.zip(dir.resolve(info.getId()), zipPath, true);
         }
     }
 
-    public void unarchive(Path dir, boolean isZipped) {
-        if(isZipped){
+    public void unarchive(Path dir, boolean isZipped) throws IOException {
+        if (isZipped) {
             int index = dir.toString().lastIndexOf(".");
             Path nodeDir = Paths.get(dir.toString().substring(0, index));
-            Utils.unzip(dir, nodeDir);
-            new AppStorageArchive(storage).unarchive(info, nodeDir);
-            Utils.deleteDirectory(new File(nodeDir.toString()));
-        }else{
+            if (Files.exists(nodeDir)) {
+                throw new FileAlreadyExistsException("Archive already exist.");
+            }
+            try {
+                Utils.unzip(dir, nodeDir);
+                new AppStorageArchive(storage).unarchive(info, nodeDir);
+            } finally {
+                Utils.deleteDirectory(new File(nodeDir.toString()));
+            }
+        } else {
             new AppStorageArchive(storage).unarchive(info, dir);
         }
     }
 
-    public void archive(Path dir){
+    public void archive(Path dir) throws IOException {
         archive(dir, false);
     }
 
-    public void unarchive(Path dir){
+    public void unarchive(Path dir) throws IOException {
         unarchive(dir, false);
     }
 }
