+++ afs-storage-api/src/main/java/com/powsybl/afs/storage/Utils.java
@@ -10,10 +10,7 @@ import com.google.common.io.ByteStreams;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
+import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.zip.ZipEntry;
@@ -27,8 +24,10 @@ import java.util.zip.ZipOutputStream;
  */
 public final class Utils {
 
+    private static final long MIN_DISK_SPACE_THRESHOLD = 10;
     private static final Logger LOGGER = LoggerFactory.getLogger(Utils.class);
-    private Utils() {
+    private Utils() throws IllegalAccessException {
+        throw new IllegalAccessException();
         //not called
     }
 
@@ -39,24 +38,47 @@ public final class Utils {
      * @param zipPath path to the zip to create
      * @throws IllegalArgumentException IllegalArgumentException
      */
-    public static void zip(Path dir, Path zipPath) throws IllegalArgumentException {
+    public static void zip(Path dir, Path zipPath, boolean deleteDirectory) throws IOException {
         try (FileOutputStream fos = new FileOutputStream(zipPath.toFile());
-             ZipOutputStream zos = new ZipOutputStream(fos)) {
+             ZipOutputStream zos = new ZipOutputStream(fos);) {
             Files.walk(dir)
                     .filter(someFileToZip -> !someFileToZip.equals(dir))
                     .forEach(
                         someFileToZip -> {
                             Path pathInZip = dir.relativize(someFileToZip);
-                            if (Files.isDirectory(someFileToZip)) {
-                                addDirectory(zos, pathInZip);
-                            } else {
-                                addFile(zos, someFileToZip, pathInZip);
+                            try {
+                                if (Files.isDirectory(someFileToZip)) {
+                                    addDirectory(zos, pathInZip);
+                                } else {
+                                    addFile(zos, someFileToZip, pathInZip);
+                                }
+                            } catch (IOException e) {
+                                throw new AfsStorageException(e.getMessage());
                             }
+
                         });
-        } catch (IOException e) {
+        } catch (IOException | AfsStorageException e) {
             LOGGER.error("The file can't be added to the zip", e);
+            throw new IOException(e);
         }
-        deleteDirectory(new File(dir.toString()));
+
+        if (deleteDirectory) {
+            deleteDirectory(new File(dir.toString()));
+        }
+    }
+
+    /**
+     * Check that there is enough space on the disk
+     * @param dir directory to save
+     * @throws IOException IOException
+     */
+    public static void checkDiskSpace(Path dir) throws IOException {
+        File archiveFile = dir.toFile();
+        long freeSpacePercent = 100 * archiveFile.getFreeSpace() / archiveFile.getTotalSpace();
+        LOGGER.info("Copying into drive with {}% free space", freeSpacePercent);
+        if (freeSpacePercent < MIN_DISK_SPACE_THRESHOLD) {
+            throw new IOException("Not enough space");
+        }
     }
 
     /**
@@ -65,7 +87,7 @@ public final class Utils {
      * @param zipPath zip Path
      * @param nodeDir path to the directory where unzip
      */
-    public static void unzip(Path zipPath, Path nodeDir) {
+    public static void unzip(Path zipPath, Path nodeDir) throws IOException {
         try (ZipInputStream zis = new ZipInputStream(new FileInputStream(zipPath.toFile()))) {
             Files.createDirectories(nodeDir);
             ZipEntry entry = zis.getNextEntry();
@@ -83,26 +105,29 @@ public final class Utils {
             zis.closeEntry();
         } catch (IOException e) {
             LOGGER.error("Unable to unzip {}", zipPath, e);
+            throw e;
         }
     }
 
-    private static void addDirectory(ZipOutputStream zos, Path relativeFilePath) {
+    private static void addDirectory(ZipOutputStream zos, Path relativeFilePath) throws IOException {
         try {
             ZipEntry entry = new ZipEntry(relativeFilePath.toString() + "/");
             zos.putNextEntry(entry);
             zos.closeEntry();
         } catch (IOException e) {
             LOGGER.error("Unable to add directory {} to zip", relativeFilePath, e);
+            throw e;
         }
     }
 
-    private static void addFile(ZipOutputStream zos, Path filePath, Path zipFilePath) {
+    private static void addFile(ZipOutputStream zos, Path filePath, Path zipFilePath) throws IOException {
         try (FileInputStream fis = new FileInputStream(filePath.toFile())) {
             ZipEntry entry = new ZipEntry(zipFilePath.toString());
             zos.putNextEntry(entry);
             ByteStreams.copy(fis, zos);
         } catch (IOException e) {
             LOGGER.error("Unable to add file {} to zip", zipFilePath, e);
+            throw e;
         }
     }
